requirements.txt
pygame-ce
win32api
win32con
win32gui